+++ src/app/add-blog/page.js
@@ -1,10 +1,15 @@
 import AddBlog from "@/components/addBlog";
+import { generateCommonMetadata } from "@/utils/helper";
 import React from "react";
-export const metadata = {
-  title: {
-    absolute: "DB Add Blog | DB Kompare",
-  },
-};
+
+const title = "Add DB Blog";
+const description =
+  "Create and publish insightful database-related content on DB Kompare. Share your expertise, tips, and experiences with database management, optimization, and migration. Whether you’re a database expert or a newcomer, the Add Blog page enables you to contribute valuable articles to help the community stay informed and make better decisions about their database systems.";
+
+export const metadata = generateCommonMetadata({
+  title,
+  description,
+});
 export default function page() {
   return <AddBlog />;
 }
